+++ src/openai.c
@@ -1474,6 +1474,7 @@ static void reportSslError(SSL *s, int ret, CONST_STRPTR where) {
  * @param proxyRequiresAuth whether the proxy requires authentication or not
  * @param proxyUsername the proxy username to use
  * @param proxyPassword the proxy password to use
+ * @param audioFormat the audio format to use
  * @return a pointer to a buffer containing the audio data or NULL -- Free it
  *with FreeVec() when you are done using it
  **/
@@ -1483,7 +1484,7 @@ APTR postTextToSpeechRequestToOpenAI(
     CONST_STRPTR openAiApiKey, ULONG *audioLength, BOOL useProxy,
     CONST_STRPTR proxyHost, UWORD proxyPort, BOOL proxyUsesSSL,
     BOOL proxyRequiresAuth, CONST_STRPTR proxyUsername,
-    CONST_STRPTR proxyPassword) {
+    CONST_STRPTR proxyPassword, AudioFormat *audioFormat) {
     struct json_object *response;
     BOOL useSSL = !useProxy || proxyUsesSSL;
 
@@ -1518,8 +1519,11 @@ APTR postTextToSpeechRequestToOpenAI(
         json_object_object_add(obj, "instructions",
                                json_object_new_string(voiceInstructions));
     }
-    json_object_object_add(obj, "response_format",
-                           json_object_new_string("pcm"));
+    json_object_object_add(
+        obj, "response_format",
+        json_object_new_string(
+            AUDIO_FORMAT_NAMES[audioFormat != NULL ? *audioFormat
+                                                   : AUDIO_FORMAT_PCM]));
     CONST_STRPTR jsonString = json_object_to_json_string(obj);
 
     STRPTR authHeader = AllocVec(256, MEMF_CLEAR | MEMF_ANY);